refactor(engine): move BoneAnimationRegistry from demo to engine layer
