+++ include/gseurat/engine/bone_animation_registry.hpp
